ShopmeWebParent/ShopmeBackEnd/src/main/java/com/deborger/shopme/admin/user/UserRepository.java
package com.deborger.shopme.admin.user;

import com.deborger.shopme.common.entity.User;
import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.CrudRepository;
import org.springframework.data.repository.query.Param;
import org.springframework.stereotype.Repository;

@Repository
public interface UserRepository extends CrudRepository<User,Integer> {

    @Query("select u from User u where u.email = :email")
    public User getUserByEmail(@Param("email") String email);
}
